+++ scripts/install_python3_dependencies.sh
@@ -6,8 +6,8 @@ mkdir -p "$FOLDER"
 sudo chown -R "$USER:$USER" "$FOLDER"
 sudo chmod -R 755 "$FOLDER"
 
-pip3 install --user --upgrade pip
-pip3 install --upgrade keyrings.alt --user
-pip3 install --user --upgrade setuptools
-pip3 install --user wheel
-pip3 install --user -r requirements.txt
+python3 -m pip install --user --upgrade pip
+python3 -m pip install --upgrade keyrings.alt --user
+python3 -m pip install --user --upgrade setuptools
+python3 -m pip install --user wheel
+python3 -m pip install --user -r requirements.txt